+++ pkg/exporter/sumologicexporter/factory_test.go
@@ -52,14 +52,13 @@ func TestCreateDefaultConfig(t *testing.T) {
 		TraceFormat: "otlp",
 
 		HTTPClientSettings: confighttp.HTTPClientSettings{
-			Timeout: 5 * time.Second,
+			Timeout: 30 * time.Second,
 			Auth: &configauth.Authentication{
 				AuthenticatorID: component.NewID("sumologic"),
 			},
 		},
-		RetrySettings:        exporterhelper.NewDefaultRetrySettings(),
-		QueueSettings:        qs,
-		DropRoutingAttribute: "",
+		RetrySettings: exporterhelper.NewDefaultRetrySettings(),
+		QueueSettings: qs,
 	})
 
 	assert.NoError(t, component.ValidateConfig(cfg))