+++ pkg/scripts_test/install_unix_test.go
@@ -85,7 +85,30 @@ func TestInstallScript(t *testing.T) {
 				checkConfigCreated,
 				checkConfigFilesOwnershipAndPermissions(rootUser, rootGroup),
 				checkUserConfigCreated,
+				checkEphemeralNotInConfig(userConfigPath),
 				checkTokenInConfig,
+				checkSystemdConfigNotCreated,
+				checkUserNotExists,
+				checkHostmetricsConfigNotCreated,
+				checkTokenEnvFileNotCreated,
+			},
+		},
+		{
+			name: "installation token and ephemeral",
+			options: installOptions{
+				skipSystemd:  true,
+				installToken: installToken,
+				ephemeral:    true,
+			},
+			preChecks: []checkFunc{checkBinaryNotCreated, checkConfigNotCreated, checkUserConfigNotCreated, checkUserNotExists},
+			postChecks: []checkFunc{
+				checkBinaryCreated,
+				checkBinaryIsRunning,
+				checkConfigCreated,
+				checkConfigFilesOwnershipAndPermissions(rootUser, rootGroup),
+				checkUserConfigCreated,
+				checkTokenInConfig,
+				checkEphemeralInConfig(userConfigPath),
 				checkSystemdConfigNotCreated,
 				checkUserNotExists,
 				checkHostmetricsConfigNotCreated,
@@ -123,6 +146,7 @@ func TestInstallScript(t *testing.T) {
 				checkBinaryCreated,
 				checkBinaryIsRunning,
 				checkConfigCreated,
+				checkRemoteConfigDirectoryNotCreated,
 				checkConfigFilesOwnershipAndPermissions(rootUser, rootGroup),
 				checkUserConfigCreated,
 				checkTokenInConfig,
@@ -130,6 +154,47 @@ func TestInstallScript(t *testing.T) {
 				checkUserNotExists,
 				checkHostmetricsConfigCreated,
 				checkHostmetricsOwnershipAndPermissions(rootUser, rootGroup),
+			},
+		},
+		{
+			name: "installation token and remotely-managed",
+			options: installOptions{
+				skipSystemd:     true,
+				installToken:    installToken,
+				remotelyManaged: true,
+			},
+			preChecks: []checkFunc{checkBinaryNotCreated, checkConfigNotCreated, checkUserConfigNotCreated, checkUserNotExists},
+			postChecks: []checkFunc{
+				checkBinaryCreated,
+				checkBinaryIsRunning,
+				checkConfigCreated,
+				checkRemoteConfigDirectoryCreated,
+				checkConfigFilesOwnershipAndPermissions(rootUser, rootGroup),
+				checkTokenInSumoConfig,
+				checkEphemeralNotInConfig(configPath),
+				checkSystemdConfigNotCreated,
+				checkUserNotExists,
+			},
+		},
+		{
+			name: "installation token, remotely-managed, and ephemeral",
+			options: installOptions{
+				skipSystemd:     true,
+				installToken:    installToken,
+				remotelyManaged: true,
+				ephemeral:       true,
+			},
+			preChecks: []checkFunc{checkBinaryNotCreated, checkConfigNotCreated, checkUserConfigNotCreated, checkUserNotExists},
+			postChecks: []checkFunc{
+				checkBinaryCreated,
+				checkBinaryIsRunning,
+				checkConfigCreated,
+				checkRemoteConfigDirectoryCreated,
+				checkConfigFilesOwnershipAndPermissions(rootUser, rootGroup),
+				checkTokenInSumoConfig,
+				checkEphemeralInConfig(configPath),
+				checkSystemdConfigNotCreated,
+				checkUserNotExists,
 			},
 		},
 		{
@@ -349,7 +414,7 @@ func TestInstallScript(t *testing.T) {
 				checkVarLogACL,
 			},
 			conditionalChecks: []condCheckFunc{checkSystemdAvailability},
-			installCode:       3, // because of invalid installation token
+			installCode:       1, // because of invalid installation token
 		},
 		{
 			name: "systemd installation token with existing user directory",
@@ -379,7 +444,7 @@ func TestInstallScript(t *testing.T) {
 				checkOutputUserAddWarnings,
 			},
 			conditionalChecks: []condCheckFunc{checkSystemdAvailability},
-			installCode:       3, // because of invalid install token
+			installCode:       1, // because of invalid install token
 		},
 		{
 			name: "systemd existing installation different token env",
@@ -409,7 +474,7 @@ func TestInstallScript(t *testing.T) {
 				checkOutputUserAddWarnings,
 			},
 			conditionalChecks: []condCheckFunc{checkSystemdAvailability},
-			installCode:       3, // because of invalid install token
+			installCode:       1, // because of invalid install token
 		},
 		{
 			name: "installation of hostmetrics in systemd during upgrade",
@@ -430,7 +495,34 @@ func TestInstallScript(t *testing.T) {
 				checkUserExists,
 				checkHostmetricsConfigCreated,
 				checkHostmetricsOwnershipAndPermissions(systemUser, systemUser),
+				checkRemoteConfigDirectoryNotCreated,
 			},
+		},
+		{
+			name: "systemd with remotely-managed",
+			options: installOptions{
+				installToken:    installToken,
+				remotelyManaged: true,
+			},
+			preChecks: []checkFunc{checkBinaryNotCreated, checkConfigNotCreated, checkUserConfigNotCreated, checkUserNotExists, checkTokenEnvFileNotCreated},
+			postChecks: []checkFunc{
+				checkBinaryCreated,
+				checkBinaryIsRunning,
+				checkConfigCreated,
+				checkRemoteConfigDirectoryCreated,
+				checkConfigFilesOwnershipAndPermissions(systemUser, systemUser),
+				checkUserConfigNotCreated,
+				checkSystemdConfigCreated,
+				checkRemoteFlagInSystemdFile,
+				checkSystemdEnvDirExists,
+				checkSystemdEnvDirPermissions,
+				checkTokenEnvFileCreated,
+				checkTokenInEnvFile,
+				checkUserExists,
+				checkVarLogACL,
+			},
+			conditionalChecks: []condCheckFunc{checkSystemdAvailability},
+			installCode:       1, // because of invalid installation token
 		},
 		{
 			name: "uninstallation without autoconfirm fails",
@@ -494,7 +586,7 @@ func TestInstallScript(t *testing.T) {
 			postChecks: []checkFunc{checkBinaryCreated, checkBinaryIsRunning, checkConfigCreated, checkDifferentTokenInConfig, checkSystemdConfigCreated,
 				checkUserExists, checkTokenEnvFileNotCreated},
 			conditionalChecks: []condCheckFunc{checkSystemdAvailability},
-			installCode:       3, // because of invalid installation token
+			installCode:       1, // because of invalid installation token
 		},
 		{
 			name: "systemd installation if token in file",